Signed comparison in example

+++ examples/algorithms/writing_new/collect_indexes__complicated_real_example.cpp
@@ -53,8 +53,9 @@ template <
 >
 void collect_indexes(R&& r, P p, std::vector<IdxType, Alloc>& res)
 {
-  EVE_ASSERT((r.end() - r.begin() <= std::numeric_limits<IdxType>::max()),
-              "The output type is not big enough to hold potential indexes");
+  EVE_ASSERT( (static_cast<std::size_t>(r.end() - r.begin()) <= std::numeric_limits<IdxType>::max())
+            , "The output type is not big enough to hold potential indexes"
+            );
 
 
   // Prepare the output in case it was not empty.
